backend/pyproject.toml
[project]
name = "multiagent-rag"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7",
  "sqlalchemy>=2.0",
  "psycopg[binary,pool]>=3.2",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "tiktoken>=0.7",
  "beautifulsoup4>=4.12",
  "python-pptx>=0.6",
  "striprtf>=0.0.26",
  "PyMuPDF>=1.24.10",
  "python-docx>=1.1.2",
  "orjson>=3.10",
  "starlette>=0.40",
  "typing-extensions>=4.12",
  "pgvector>=0.4.1",
  "debugpy>=1.8.17",

  # --- Added for robust parsing/OCR fallback ---
  "pillow>=10.4",
  "pytesseract>=0.3.13",
  "lxml>=5.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.10",
]

[tool.uvicorn]
reload = true
